feat(signup): implements feedback
- create a signup routes
- create a signup controller
- write a test for the signup

[Finishes AC-14]
