+++ src/routes/index.js
@@ -1,10 +1,10 @@
 import express from 'express';
 import user from '../users/route';
-import account from '../accounts/route';
+import signup from '../signup/route';
 
 const app = express();
 
 app.use(user);
-app.use(account);
+app.use(signup);
 
 export default app;
