+++ src/signup/route.js
@@ -1,0 +1,10 @@
+import express from 'express';
+
+import signupValidation from '../middlewares/signup';
+import SignupController from '.';
+
+const router = express.Router();
+
+router.post('/signup', signupValidation.registerAccountValidation, SignupController.registerAccount);
+
+export default router;
